+++ src/main/java/com/SchoolManagementSystem/repository/StudentRepository.java
@@ -3,5 +3,5 @@ package com.SchoolManagementSystem.repository;
 import com.SchoolManagementSystem.entity.Student;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface StudentRepository extends JpaRepository<Student,Integer>{
+public interface StudentRepository extends JpaRepository<Student,Long>{
 }
