+++ src/main/java/com/SchoolManagementSystem/service/Impl/StudentServiceImpl.java
@@ -13,13 +13,13 @@ public class StudentServiceImpl implements StudentService {
     StudentRepository studentRepository;
 
     @Override
-    public Student getStudent(Integer studentId) {
+    public Student getStudent(Long studentId) {
         Student student = studentRepository.findById(studentId).orElseThrow(() -> new RuntimeException("There is no such student with that ID"));
         return student;
     }
 
     @Override
-    public void deleteStudent(Integer studentId) {
+    public void deleteStudent(Long studentId) {
         studentRepository.deleteById(studentId);
     }
