+++ src/main/java/com/SchoolManagementSystem/service/StudentService.java
@@ -6,7 +6,7 @@ import com.SchoolManagementSystem.entity.Student;
 public interface StudentService {
     Student createStudent(Student student);
     Student updateStudent(Student student);
-    Student getStudent(Integer studentId);
-    void deleteStudent(Integer studentId);
+    Student getStudent(Long studentId);
+    void deleteStudent(Long studentId);
 
 }
